% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{dd_bin_edges}
\alias{dd_bin_edges}
\title{Construct bin edges spanning a range}
\usage{
dd_bin_edges(op_range, bins = 20L, inc = NULL)
}
\description{
Construct bin edges spanning a range
}
\keyword{internal}
