% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{dd_bin_index}
\alias{dd_bin_index}
\title{Assign values to half-open bins \verb{[edges[i], edges[i+1])}, clamped to range}
\usage{
dd_bin_index(x, edges)
}
\description{
Assign values to half-open bins \verb{[edges[i], edges[i+1])}, clamped to range
}
\keyword{internal}
