% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_fit.R
\name{dd_cv_error}
\alias{dd_cv_error}
\title{K-fold cross-validation error for a polynomial fit}
\usage{
dd_cv_error(x, y, degree, threshold, alpha = 0, folds = 5)
}
\description{
K-fold cross-validation error for a polynomial fit
}
\keyword{internal}
