% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dd_diffusion}
\alias{dd_diffusion}
\title{Binned diffusion estimate}
\usage{
dd_diffusion(dd)
}
\arguments{
\item{dd}{A \code{"daddy"} object from \code{\link[=dd_analyse]{dd_analyse()}}.}
}
\value{
A data frame with columns \code{op}, \code{avg_diff}, \code{diff_se}, \code{n}.
}
\description{
Binned diffusion estimate
}
