% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dd_drift}
\alias{dd_drift}
\title{Binned drift estimate}
\usage{
dd_drift(dd)
}
\arguments{
\item{dd}{A \code{"daddy"} object from \code{\link[=dd_analyse]{dd_analyse()}}.}
}
\value{
A data frame with columns \code{op}, \code{avg_drift}, \code{drift_se}, \code{n}.
}
\description{
Binned drift estimate
}
