% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_fit.R
\name{dd_poly_dictionary}
\alias{dd_poly_dictionary}
\title{Build a polynomial design matrix \verb{[1, x, x^2, ..., x^degree]}}
\usage{
dd_poly_dictionary(x, degree)
}
\description{
Build a polynomial design matrix \verb{[1, x, x^2, ..., x^degree]}
}
\keyword{internal}
